dice-roller/pyproject.toml
[project]
name = "dice-roller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "uvicorn>=0.37.0",
    "opentelemetry-distro>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
]
